requirements.txt
ipykernel
transformers
langchain==0.3.15
langchain-groq==0.3.8
python-dotenv
pandas
streamlit
matplotlib